requirements.txt
discord.py>=2.3.0
discord-ext-voice-recv>=0.3.0a120
python-dotenv>=1.0.0
mcrcon>=0.1.0
faster-whisper>=1.0.0
pydub>=0.25.1
numpy>=1.24.0
pytest>=7.0.0
pytest-cov>=4.0.0
